+++ src/routes/route.jsx
@@ -1,7 +1,7 @@
 import Home from '@/pages/home';
 import ProductDetail from '@/pages/productDetail';
 import Signup from '@/pages/signup';
-import ProductCreate from '@/pages/product/productCreate';
+import ProductCreate from '@/pages/productCreate';
 import { createBrowserRouter } from 'react-router-dom';
 
 const router = createBrowserRouter([
